+++ src/main/java/com/qcloud/cos/demo/BucketInventoryDemo.java
@@ -9,6 +9,7 @@ import com.qcloud.cos.model.ListBucketInventoryConfigurationsRequest;
 import com.qcloud.cos.model.ListBucketInventoryConfigurationsResult;
 import com.qcloud.cos.model.SetBucketInventoryConfigurationRequest;
 import com.qcloud.cos.model.DeleteBucketInventoryConfigurationRequest;
+import com.qcloud.cos.model.inventory.PostBucketInventoryConfigurationResult;
 import com.qcloud.cos.model.inventory.InventoryConfiguration;
 import com.qcloud.cos.model.inventory.InventoryCosBucketDestination;
 import com.qcloud.cos.model.inventory.InventoryFrequency;
@@ -17,30 +18,46 @@ import com.qcloud.cos.model.inventory.InventoryFormat;
 import com.qcloud.cos.model.inventory.InventoryDestination;
 import com.qcloud.cos.model.inventory.InventorySchedule;
 import com.qcloud.cos.model.inventory.InventoryPrefixPredicate;
+import com.qcloud.cos.model.inventory.InventoryAndPredicate;
 import com.qcloud.cos.model.inventory.InventoryFilter;
 import com.qcloud.cos.model.inventory.InventoryIncludedObjectVersions;
 import com.qcloud.cos.model.inventory.InventoryOptionalField;
+import com.qcloud.cos.model.Tag.Tag;
 import com.qcloud.cos.region.Region;
 
+import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
 public class BucketInventoryDemo {
+    private static String secretId = System.getenv("SECRETID");
+    private static String secretKey = System.getenv("SECRETKEY");
+    private static String bucketName = System.getenv("BUCKET_NAME");
+    private static String region = System.getenv("REGION");
+    private static COSClient cosClient = createCli();
+
     public static void main(String[] args) {
-        setGetDeleteBucketInventoryDemo();
-        setBucketInventoryDemo();
+        try {
+            setGetDeleteBucketInventoryDemo();
+            setBucketInventoryDemo();
+//            postBucketInventoryDemo();
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            cosClient.shutdown();
+        }
+    }
+
+    private static COSClient createCli() {
+        // 初始化用户身份信息(secretId, secretKey)
+        COSCredentials cred = new BasicCOSCredentials(secretId,secretKey);
+        // 设置bucket的区域, COS地域的简称请参照 https://www.qcloud.com/document/product/436/6224
+        ClientConfig clientConfig = new ClientConfig(new Region(region));
+        // 生成cos客户端
+        return new COSClient(cred, clientConfig);
     }
 
     private static void setGetDeleteBucketInventoryDemo() {
-        // 1 初始化用户身份信息(secretId, secretKey)
-        COSCredentials cred = new BasicCOSCredentials("AKIDXXXXXXXX", "1A2Z3YYYYYYYYYY");
-        // 2 设置bucket的区域, COS地域的简称请参照 https://www.qcloud.com/document/product/436/6224
-        ClientConfig clientConfig = new ClientConfig(new Region("ap-guangzhou"));
-        // 3 生成cos客户端
-        COSClient cosclient = new COSClient(cred, clientConfig);
-        // bucket名需包含appid
-        String bucketName = "mybucket-12500000000";
-
         InventoryConfiguration inventoryConfiguration = new InventoryConfiguration();
         InventoryCosBucketDestination inventoryCosBucketDestination = new InventoryCosBucketDestination();
         // 设置清单的输出目标存储桶的格式和前缀等
@@ -70,25 +87,25 @@ public class BucketInventoryDemo {
         SetBucketInventoryConfigurationRequest setBucketInventoryConfigurationRequest = new SetBucketInventoryConfigurationRequest();
         setBucketInventoryConfigurationRequest.setBucketName(bucketName);
         setBucketInventoryConfigurationRequest.setInventoryConfiguration(inventoryConfiguration);
-        cosclient.setBucketInventoryConfiguration(setBucketInventoryConfigurationRequest);
+        cosClient.setBucketInventoryConfiguration(setBucketInventoryConfigurationRequest);
 
         inventoryConfiguration.setId("2");
         inventorySchedule.setFrequency(InventoryFrequency.Weekly);
-        cosclient.setBucketInventoryConfiguration(setBucketInventoryConfigurationRequest);
+        cosClient.setBucketInventoryConfiguration(setBucketInventoryConfigurationRequest);
 
         // 获取指定id的清单配置
-        GetBucketInventoryConfigurationResult getBucketInventoryConfigurationResult = cosclient.getBucketInventoryConfiguration(bucketName, "1");
+        GetBucketInventoryConfigurationResult getBucketInventoryConfigurationResult = cosClient.getBucketInventoryConfiguration(bucketName, "1");
 
         // 批量获取清单
         ListBucketInventoryConfigurationsRequest listBucketInventoryConfigurationsRequest = new ListBucketInventoryConfigurationsRequest();
         listBucketInventoryConfigurationsRequest.setBucketName(bucketName);
-        ListBucketInventoryConfigurationsResult listBucketInventoryConfigurationsResult = cosclient.listBucketInventoryConfigurations(listBucketInventoryConfigurationsRequest);
+        ListBucketInventoryConfigurationsResult listBucketInventoryConfigurationsResult = cosClient.listBucketInventoryConfigurations(listBucketInventoryConfigurationsRequest);
 
         // 删除指定清单
         DeleteBucketInventoryConfigurationRequest deleteBucketInventoryConfigurationRequest = new DeleteBucketInventoryConfigurationRequest();
         deleteBucketInventoryConfigurationRequest.setBucketName(bucketName);
         deleteBucketInventoryConfigurationRequest.setId("1");
-        cosclient.deleteBucketInventoryConfiguration(deleteBucketInventoryConfigurationRequest);
+        cosClient.deleteBucketInventoryConfiguration(deleteBucketInventoryConfigurationRequest);
     }
 
     private static void setBucketInventoryDemo() {
@@ -143,5 +160,48 @@ public class BucketInventoryDemo {
 
         cosclient.setBucketInventoryConfiguration(request);
         cosclient.shutdown();
+    }
+
+    private static void postBucketInventoryDemo() {
+        InventoryConfiguration inventoryConfiguration = new InventoryConfiguration();
+        InventoryCosBucketDestination inventoryCosBucketDestination = new InventoryCosBucketDestination();
+        // 设置清单的输出目标存储桶的格式和前缀等
+        inventoryCosBucketDestination.setAccountId("100000000001");
+        inventoryCosBucketDestination.setBucketArn("qcs::cos:ap-guangzhou::mybucket-12500000000");
+        inventoryCosBucketDestination.setEncryption(new ServerSideEncryptionCOS());
+        inventoryCosBucketDestination.setFormat(InventoryFormat.CSV);
+        inventoryCosBucketDestination.setPrefix("inventory-output");
+        InventoryDestination inventoryDestination = new InventoryDestination();
+        inventoryDestination.setCosBucketDestination(inventoryCosBucketDestination);
+        inventoryConfiguration.setDestination(inventoryDestination);
+
+        // 设置清单，扫描前缀和id等
+        inventoryConfiguration.setId("1");
+        // prefix filter
+        InventoryPrefixPredicate inventoryFilter = new InventoryPrefixPredicate("test/");
+        inventoryConfiguration.setInventoryFilter(new InventoryFilter(inventoryFilter));
+
+        // and Filter
+//        InventoryAndPredicate inventoryAndPredicate = new InventoryAndPredicate();
+//        inventoryAndPredicate.setPrefix("bbb");
+//        List<Tag> filterTags = new ArrayList<>();
+//        filterTags.add(new Tag("Testkey1", "value1"));
+//        filterTags.add(new Tag("Testkey2", "value2"));
+//        inventoryAndPredicate.setTags(filterTags);
+//        inventoryConfiguration.setInventoryFilter(new InventoryFilter(inventoryAndPredicate));
+
+        inventoryConfiguration.setIncludedObjectVersions(InventoryIncludedObjectVersions.All);
+        // 设置可选的输出字段
+        List<String> optionalFields = new LinkedList<String>();
+        optionalFields.add(InventoryOptionalField.Size.toString());
+        optionalFields.add(InventoryOptionalField.LastModifiedDate.toString());
+        optionalFields.add(InventoryOptionalField.Tag.toString());
+        inventoryConfiguration.setOptionalFields(optionalFields);
+
+        SetBucketInventoryConfigurationRequest setBucketInventoryConfigurationRequest = new SetBucketInventoryConfigurationRequest();
+        setBucketInventoryConfigurationRequest.setBucketName(bucketName);
+        setBucketInventoryConfigurationRequest.setInventoryConfiguration(inventoryConfiguration);
+        PostBucketInventoryConfigurationResult result = cosClient.postBucketInventoryConfiguration(setBucketInventoryConfigurationRequest);
+        System.out.println("finish post bucket inventory, jobId: " + result.getJobId() + ", reqId: " + result.getRequestId());
     }
 }