+++ src/main/java/com/qcloud/cos/internal/RequestXmlFactory.java
@@ -18,6 +18,10 @@
 
 package com.qcloud.cos.internal;
 
+import static com.qcloud.cos.model.ciModel.xml.CIMediaXmlFactory.addAigcMetadata;
+import static com.qcloud.cos.model.ciModel.xml.CIMediaXmlFactory.addCallBackKafkaConfig;
+import static com.qcloud.cos.model.ciModel.xml.CIMediaXmlFactory.addCallBackMqConfig;
+
 import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Field;
 import java.util.Collections;
@@ -632,6 +636,7 @@ public class RequestXmlFactory {
             addIfNotNull(xml, "ResoAdjMethod", request.getTransConfig().getResoAdjMethod());
             addIfNotNull(xml, "TransMode", request.getTransConfig().getTransMode());
             addIfNotNull(xml, "VideoBitrateAdjMethod", request.getTransConfig().getVideoBitrateAdjMethod());
+            addAigcMetadata(xml, request.getTransConfig().getAigcMetadata());
             xml.end();
 
             addVideo(xml, request);
@@ -690,6 +695,11 @@ public class RequestXmlFactory {
         xml.start("Input");
         xml.start("Object").value(docJobObject.getInput().getObject()).end();
         xml.end();
+        addIfNotNull(xml, "CallBack", docJobObject.getCallBack());
+        addIfNotNull(xml, "CallBackFormat", docJobObject.getCallBackFormat());
+        addIfNotNull(xml, "CallBackType", docJobObject.getCallBackType());
+        addCallBackMqConfig(xml, docJobObject.getCallBackMqConfig());
+        addCallBackKafkaConfig(xml, docJobObject.getCallBackKafkaConfig());
 
         if (CheckObjectUtils.objIsNotValid(docJobObject)){
             xml.start("Operation");
@@ -716,7 +726,7 @@ public class RequestXmlFactory {
             addIfNotNull(xml, "PicPagination", docProcess.getPicPagination());
             addIfNotNull(xml, "ImageDpi", docProcess.getImageDpi());
             xml.end();
-
+            addIfNotNull(xml, "UserData", docJobObject.getOperation().getUserData());
             xml.end();
         }
         xml.end();