+++ src/main/java/com/qcloud/cos/model/ciModel/job/MediaListJobResponse.java
@@ -1,6 +1,8 @@
 package com.qcloud.cos.model.ciModel.job;
 
 import com.qcloud.cos.model.CiServiceResult;
+import com.thoughtworks.xstream.annotations.XStreamAlias;
+import com.thoughtworks.xstream.annotations.XStreamImplicit;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,9 +11,12 @@ import java.util.List;
 /**
  * 媒体处理 任务列表响应实体 https://cloud.tencent.com/document/product/460/48234
  */
+@XStreamAlias("Response")
 public class MediaListJobResponse extends CiServiceResult {
+    @XStreamImplicit(itemFieldName = "JobsDetail")
     private List<MediaJobObject> jobsDetailList;
 
+    @XStreamAlias("NextToken")
     private String nextToken;
 
     public List<MediaJobObject> getJobsDetailList() {